+++ taskbar.c
@@ -7,11 +7,6 @@
 #include "window.h"
 #include "event.h"
 
-static GC gc_taskbar;
-static GC gc_taskbar_button;
-static GC gc_taskbar_button_focus;
-static GC gc_taskbar_button_hidden;
-static GC gc_taskbar_button_border;
 taskbar_t *tb;
 
 static Pixmap pixmap;
@@ -25,32 +20,6 @@ taskbar_t* taskbar_init() {
     tb->height = screen_height;
     tb->tb_buttons = malloc(sizeof(taskbar_button_h));
     tb->buttons_length = 0;
-
-    XColor color_taskbar = { .red = 65535, .green = 65535, .blue = 65535 };
-    XAllocColor(display, colormap, &color_taskbar);
-    XGCValues gcv_taskbar = { .foreground = color_taskbar.pixel };
-    gc_taskbar = XCreateGC(display, window, GCForeground, &gcv_taskbar);
-
-    XColor color_taskbar_button = { .red = 65535, .green = 35000, .blue = 10000 };
-    XAllocColor(display, colormap, &color_taskbar_button);
-    XGCValues gcv_taskbar_button = { .foreground = color_taskbar_button.pixel };
-    gc_taskbar_button = XCreateGC(display, window, GCForeground, &gcv_taskbar_button);
-
-	XColor color_taskbar_button_focus = { .red = 20000, .green = 50000, .blue = 20000 };
-    XAllocColor(display, colormap, &color_taskbar_button_focus);
-    XGCValues gcv_taskbar_button_focus = { .foreground = color_taskbar_button_focus.pixel };
-    gc_taskbar_button_focus = XCreateGC(display, window, GCForeground, &gcv_taskbar_button_focus);
-
-    XColor color_taskbar_button_hidden = { .red = 20000, .green = 20000, .blue = 20000 };
-    XAllocColor(display, colormap, &color_taskbar_button_hidden);
-    XGCValues gcv_taskbar_button_hidden = { .foreground = color_taskbar_button_hidden.pixel };
-    gc_taskbar_button_hidden = XCreateGC(display, window, GCForeground, &gcv_taskbar_button_hidden);
-
-    XGCValues gcv_taskbar_button_border = {
-        .foreground = color_taskbar.pixel,
-        .background = black_pixel
-    };
-    gc_taskbar_button_border = XCreateGC(display, window, GCForeground | GCBackground, &gcv_taskbar_button_border);
 
 	pixmap = XCreatePixmap(display, window, tb->width, tb->height, XDefaultDepthOfScreen(screen));